# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ActivityItem {
  id: string;
  name: string;
  category: string;
  proxy: string;
  proxyYear: string;
  unit: string;
  unitLabel: string;
  valuePerUnit: number;
  sdg: string;
  sdgColor: string;
  description: string;
}

export interface ActivitiesResponse {
  activities: ActivityItem[];
  categories: string[];
}

export interface SelectedActivity {
  activityId: string;
  quantity: number;
  hoursPerYear: number;
  description?: string;
}

export interface ImpactInput {
  description: string;
  activities: SelectedActivity[];
  donationsGBP: number;
  additionalVolunteerHours: number;
}

export interface ActivityBreakdown {
  activityId: string;
  activityName: string;
  category: string;
  sdg: string;
  sdgColor: string;
  impactValue: number;
  hours: number;
}

export type ImpactResultSdgBreakdownsItem = {
  sdg: string;
  sdgColor: string;
  value: number;
};

export type ImpactResultExplanations = {
  impact: string;
  contribution: string;
  donations: string;
  personalDevelopment: string;
};

export interface ImpactResult {
  totalValue: number;
  impactValue: number;
  contributionValue: number;
  donationsValue: number;
  personalDevelopmentValue: number;
  totalHours: number;
  activityBreakdowns: ActivityBreakdown[];
  sdgBreakdowns: ImpactResultSdgBreakdownsItem[];
  explanations: ImpactResultExplanations;
}

export interface SaveImpactInput {
  userId: string;
  name: string;
  impactResult: ImpactResult;
  activities: SelectedActivity[];
  donationsGBP: number;
  additionalVolunteerHours: number;
}

export interface SavedImpact {
  id: string;
  userId: string;
  name: string;
  createdAt: string;
  impactResult: ImpactResult;
}

export interface ImpactHistoryResponse {
  records: SavedImpact[];
}

export interface SuggestionsInput {
  currentActivities: string[];
  interests?: string[];
  availableHoursPerWeek: number;
}

export interface SuggestionItem {
  activityId: string;
  activityName: string;
  category: string;
  sdg: string;
  sdgColor: string;
  reason: string;
  estimatedImpactPerYear: number;
  recommendedHoursPerWeek: number;
}

export interface SuggestionsResponse {
  suggestions: SuggestionItem[];
}

export type GetImpactHistoryParams = {
  userId: string;
};