# lib/api-zod/src/generated/types/activitiesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from "./activityItem";

export interface ActivitiesResponse {
  activities: ActivityItem[];
  categories: string[];
}