# lib/api-zod/src/generated/types/activityBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityBreakdown {
  activityId: string;
  activityName: string;
  category: string;
  sdg: string;
  sdgColor: string;
  impactValue: number;
  hours: number;
}