# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  name: string;
  category: string;
  proxy: string;
  proxyYear: string;
  unit: string;
  unitLabel: string;
  valuePerUnit: number;
  sdg: string;
  sdgColor: string;
  description: string;
}