# lib/api-zod/src/generated/types/impactHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { SavedImpact } from "./savedImpact";

export interface ImpactHistoryResponse {
  records: SavedImpact[];
}