# lib/api-zod/src/generated/types/impactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { SelectedActivity } from "./selectedActivity";

export interface ImpactInput {
  description: string;
  activities: SelectedActivity[];
  donationsGBP: number;
  additionalVolunteerHours: number;
}