# lib/api-zod/src/generated/types/impactResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityBreakdown } from "./activityBreakdown";
import type { ImpactResultExplanations } from "./impactResultExplanations";
import type { ImpactResultSdgBreakdownsItem } from "./impactResultSdgBreakdownsItem";

export interface ImpactResult {
  totalValue: number;
  impactValue: number;
  contributionValue: number;
  donationsValue: number;
  personalDevelopmentValue: number;
  totalHours: number;
  activityBreakdowns: ActivityBreakdown[];
  sdgBreakdowns: ImpactResultSdgBreakdownsItem[];
  explanations: ImpactResultExplanations;
}