# lib/api-zod/src/generated/types/saveImpactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { ImpactResult } from "./impactResult";
import type { SelectedActivity } from "./selectedActivity";

export interface SaveImpactInput {
  userId: string;
  name: string;
  impactResult: ImpactResult;
  activities: SelectedActivity[];
  donationsGBP: number;
  additionalVolunteerHours: number;
}