# lib/api-zod/src/generated/types/savedImpact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */
import type { ImpactResult } from "./impactResult";

export interface SavedImpact {
  id: string;
  userId: string;
  name: string;
  createdAt: string;
  impactResult: ImpactResult;
}