# lib/api-zod/src/generated/types/selectedActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */

export interface SelectedActivity {
  activityId: string;
  quantity: number;
  hoursPerYear: number;
  description?: string;
}