# lib/api-zod/src/generated/types/suggestionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestionItem {
  activityId: string;
  activityName: string;
  category: string;
  sdg: string;
  sdgColor: string;
  reason: string;
  estimatedImpactPerYear: number;
  recommendedHoursPerWeek: number;
}