# lib/api-zod/src/generated/types/suggestionsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Impact API
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestionsInput {
  currentActivities: string[];
  interests?: string[];
  availableHoursPerWeek: number;
}